tweetNsell-server/requirements.txt
Django==2.2.3
djangorestframework==3.10.1
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
django-cors-headers==2.5.2
twitter==1.18.0
python-dotenv==0.10.3
